requirements.txt
# AI 生成
openai>=1.30.0

# 浏览器自动化
playwright>=1.44.0

# 图片处理
Pillow>=10.3.0

# HTTP 请求
httpx>=0.27.0

# 配置管理
python-dotenv>=1.0.0

# 日志
loguru>=0.7.2
